fixed infinite sync issue, converted InternalDocHandle into a class, only accessible inside the repo actor
